fix: Add SYNC permission definition (#77)
If the Ping app is installed without the Gateway app, and the Gateway app is later installed, a security exception will be thrown from the Ping app for missing a permission. That's because the permission was just defined on the Gateway app, and non-existent at the time of installation.

Defining it in both places should fix the issue.